+++ tcob/src/audio/SpeechGenerator.cpp
@@ -18,8 +18,8 @@ auto speech_generator::create_buffer [[nodiscard]] (std::string const& text) -> 
     auto* data {speech_gen(&frames, text.c_str(), nullptr)};
 
     std::vector<f32> databuf;
-    databuf.resize(frames * channels);
-    for (i32 i {0}; i < frames * channels; ++i) { databuf[i] = data[i] / 32768.0f; }
+    databuf.resize(static_cast<usize>(frames * channels));
+    for (i32 i {0}; i < frames * channels; ++i) { databuf[static_cast<usize>(i)] = data[i] / 32768.0f; }
     speech_free(data, nullptr);
 
     return buffer::Create({.Channels = channels, .SampleRate = sampleRate, .FrameCount = frames}, databuf);
